Fix test case: generic_config_updater/test_ipv6.py::test_ipv6_neighbor_admin_change (#5647)
This case fails every time on our testbeds because it tries to replace the admin_status value of a bgp neighour, while the admin_status field doesn't exist in the default running configuration generated from minigraph(the admin_status is up without a explicit "admin_status": "up" config in db).
To fix this, need to add this config before replace it.
And add a topology mark for the test according to the description in [GCU] Add config ipv6 BGP neighbor test #5061 .

- What is the motivation for this PR?
Ensure that the configuration exists in the running config before it is replaced by the patch .
Add a topology mark for the test to ensure it runs with the correct topologies. According to the description in [GCU] Add config ipv6 BGP neighbor test #5061, this test support t0 and t1 testbeds.

- How did you do it?
Modified the patch to add this config before replace it. If the config already exists before the adding, it won't take any effects because the config is duplicated.
Add the topology mark.

- How did you verify/test it?
Run this test case on testbeds with and without the admin_status field in running configs, all passed.