Add a tool to check testbed and inventory file (#5676)
What is the motivation for this PR?
When we may make some mistakes while adding or updating testbed
configurations in testbed file and inventory files. These mistakes
are hard to find and can cause many troubles.

How did you do it?
This PR added a tool to do some basic checks of testbed file and the
related inventory files. This tool only supports the yaml format testbed file.

Checked items:
* Check if the topologies defined in vm inventory file are valid.
* Check if there is duplicated testbed name in testbed file.
* For each test server, check if there is duplicated group name.
* Check if the topo name of each testbed is valid.
* Check if the PTF host of each testbed is defined in inventory
  file. Also check if the PTF IP in testbed file matches the IP
  address defined in inventory file.
* Check if there is duplicated IP address in inventory files.

The tool will print out results of all the failed checks.

Example usage:

    $ python ansible/check_testbed_and_inventory_file.py -t vtestbed.yaml -m veos_vtb

Another fix in this PR is to correct inconsistent PTF IP of the
dualtor-mixed testbed. This issue was indeed found by this tool.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>