Fix test case: generic_config_updater/test_eth_interface.py::test_replace_mtu (#5692)
The test_replace_mtu case uses Ethernet0 to do the replacing mtu test. But in some testbeds, Ethernet0 is a member of a port channel. The mtu of port channel member should not be modified directly.

- What is the motivation for this PR?
The test_replace_mtu case uses Ethernet0 to do the replacing mtu test. But in some testbeds, Ethernet0 is a member of a port channel. The mtu of port channel member should not be modified directly(can't config mtu of a port channel member via CLI).

- How did you do it?
Add get_ethernet_port_not_in_portchannel(duthost) to select a ethernet port not in a port channel for the test.
Modify check_interface_status(duthost, field, interface='Ethernet0') to accept a interface name parameter and able to check the specified interface, keeping 'Ethernet0' as the default value for other cases.
Use get_ethernet_port_not_in_portchannel in mtu test to get a port not in a port channel.

- How did you verify/test it?
Have run this case on t0-64 and t-64-lag testbeds, all passed.