Modify minigraph_dpg_asic template to generate unique port channel (#5705)
What is the motivation for this PR?
snmp_interfaces test case fails on supervisor on packet chassis because the internal Portchannel name generated is not unique for the supervisor. The portchannel names generated is unique within a namespace but not across namespaces. SNMP expects the portchannel names to unique as the port channel index used in snmp output has to be unique.

How did you do it?
Modify dpg_asic template to generate unique port channel index for supervisor node with maximum of 16 asics.

How did you verify/test it?
Tested on multi-asic platform and linecard to ensure unique port channel indices are generated.