Merged PR 548: Add support of using different agent pool for testbed pipelines
Add support of using different agent pool for testbed pipelines

To run the testbed pipelines on testbeds in SVC lab, it would be better
to use the agent pool running in SVC lab server.

This change added parameter for some of the testbed pipelines to
support selecting different agent pool.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>