Remove dafault acl tables before testing subinterfaces based on portchannels
In sub interfaces test, the portchannel cases add ethernet ports to portchannels, but on some testbeds, the selected ports are already in the default acl tables, so they can not be added to a port channel.
Remove the default acl tables before the port channel cases, and the acl tables will be added back automatically by config reload in single test case and test suite teardowns.