catch exception (#5738)
Description of PR
Only scapy version over 2.4.5 supports macsec. When we run test using scapy which version is less than 2.4.5, it will raise ImportError and cause the test failure. In this pr, we add exception handling for macsec ImportError, so, it will not break the test.

What is the motivation for this PR?
Only scapy version over 2.4.5 supports macsec. When we run test using scapy which version is less than 2.4.5, it will raise ImportError and cause the test failure. In this pr, we add exception handling for macsec ImportError, so, it will not break the test.

How did you do it?
Add try...except to catch the error.

How did you verify/test it?
By running the test and it will not break the test.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>