Update test_config_reload.py to support multi-asic platforms
test_reload_configuration_checks rebooted the DUT and waited for only 5 seconds before doing 'config reload -y' and
checking for 'Retry later' in the output.

However, on a multi-asic platform, the 'config reload -y' command would through an exception with backtrace:
   "  File \"/usr/local/bin/config\", line 5, in <module>",
   "    from config.main import config",
   "  File \"/usr/local/lib/python3.9/dist-packages/config/main.py\", line 28, in <module>",
   "    from utilities_common import bgp_util",
   "  File \"/usr/local/lib/python3.9/dist-packages/utilities_common/bgp_util.py\", line 6, in <module>",
   "    import utilities_common.cli as clicommon",
   "  File \"/usr/local/lib/python3.9/dist-packages/utilities_common/cli.py\", line 187, in <module>",
   "    iface_alias_converter = InterfaceAliasConverter()",
   "  File \"/usr/local/lib/python3.9/dist-packages/utilities_common/cli.py\", line 124, in __init__",
   "    load_db_config()",
   "  File \"/usr/local/lib/python3.9/dist-packages/utilities_common/general.py\", line 30, in load_db_config",
   "    swsscommon.SonicDBConfig.load_sonic_global_db_config()",
   "  File \"/usr/lib/python3/dist-packages/swsscommon/swsscommon.py\", line 1267, in load_sonic_global_db_config",
   "    SonicDBConfig.initializeGlobalConfig(global_db_file_path)",
   "  File \"/usr/lib/python3/dist-packages/swsscommon/swsscommon.py\", line 1262, in initializeGlobalConfig",
   "    return _swsscommon.SonicDBConfig_initializeGlobalConfig(*args, **kwargs)",
   "RuntimeError: Sonic database config file syntax error >> Sonic database config file doesn't exist at /var/run/redis/sonic-db/../../redis0/sonic-db/database_config.json"

Reason for the above is that the multi-asic database dockers are not all initialized/started within the 5 seconds after reboot and able to ssh login.
Therefore, added checks if all database containers have started before issuing config reload command.

The other issue was when trying to stop swss service, we were not taking care of multi-asic swss instances.
Therefore, if duthost is multi-asic, we stop swss on all asics, otherwise stop swss on one ASIC.