Merged PR 553: Improve nightly test to support BSL test
This change improved the nightly test template to support BSL testing.

A new parameter `RUN_BSL_TEST` was added to the nightly test template.
For the testbeds that need to run BSL testing, need to pass down True to this parameter while calling the nightly test template. Then nightly test will:
* set the switch to l2 mode after the run tests step.
* run pytest with `-m bsl` to run the scripts with `bsl` mark.
* restore the switch to l3 mode using `testbed-cli.sh deploy-mg`.

Future improvements:
The scripts to be executed in l2 mode may need to be fine tuned or fixed.