Fix container_autorestart test case to run on chassis (#5478)
What is the motivation for this PR?
Current issues while running container autorestart test case on chassis:
Test case uses enum_rand_one_per_hwsku_frontend_hostname, so test case does not run on supervisor and only runs on linecard.
Test case uses enum_dut_feature_container, which iterates over a list of container names generated for the testbed. This fixture is a combination of container name and dut name. If the dutname in this fixture does not match with the randomly selected dut, then the test is skipped. This is not efficient and can end up skipping certain containers for multiple sku testbed setup.
Test case checks for BGP sessions that are up before and after the test case. This check fails for multi-asic line card, because, in multiasic linecard, BGP neighbor IP can be same for different asics in the same linecard (loopback IP of neighbor asic).
How did you do it?
Modify to use enum_rand_one_per_hwsku_hostname, so that the test case runs on sup and a randomly selected linecard.
With the above change, made sure that the config reload is done on all the selected DUTs after the test is complete.
Instead of using the container name and skipping tests if the dut name in the fixture does not match with the selected dut, use enum_dut_feature and enum_rand_one_asic_index so that the test is executed on all features for the selected DUT.
Modify check BGP sessions, to get BGP sessions for each asic namespace, and check for each asic namespace separately.
How did you verify/test it?
Ran test on chassis with multi-asic Line card.
Ensure that 'autorestart' is enabled for all features.
Else when teamd container goes down, if swss autorestart is not enabled, then teamd docker will not come up and further tests on swss syncd will also fail. issue logged: #5601