[python3] Refact `icmp_responder` (#5696)
Approach
What is the motivation for this PR?
Migrate icmp_responder to Python3.
Improve icmp_responder performance.

How did you do it?
Previously, icmp_responder was running in a thread pool, for each port, there is a select event loop to poll the packet socket(binding to the port) for any ICMP request and respond to that request.
With the presence of GIL, the Python interpreter could only respond to one port read/write operations pair once a time. The threaded icmp_responder will not help much here but introduce overhead in the thread context switching. And
With the introduction of asyncio, both read/write operations to the packet socket could be asynchronous.

How did you verify/test it?