Parse added timestamp of testcases and add kusto support
What is the motivation for this PR?
Timestamp of testcases has been added and will be shown in tr.xml, and
it's necessary to show that in kusto

How did you do it?
In parser, I consider timestamp as testcase attributes, so I parse that
with testcase attributes like classname, file..., not testcase
properties
In kusto, I added 2 columns StartTime and EndTime for timestamp

How did you verify/test it?
Run a test.

Any platform specific information?
Supported testbed topology if it's a new test case?