Fix a bug in `tests/common/plugins/conditional_mark/tests_mark_conditions.yaml` (#5830)
Description of PR
There is an error due to the lack of a quotation in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. Add the quotation in this pr.

Summary:
Fixes #5806

What is the motivation for this PR?
There is an error due to the lack of a quotation in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. Add the quotation in this pr.

How did you do it?
Add the lack quotation.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>