Fixes for test_reboot and deleting iptables rule for platform api
test_reboot:
 - Some vendors don't support rebooting of modules other than
themselves. For this scenario, we check whether the DUT supports
rebooting other modules from 'reboot_other_modules' from platform.json
and if not supported, then skip rebooting the other modules

conftest:
 - Sometimes we see failure in deleting the iptable rule for allowing
incoming traffic for the platform api http server as the rule is not
present. For this, added check to see if the rule exists, and delete it
then only