Get token from environment variable for accessing Github (#5767)
Description of PR
We can use github issues as conditions in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. If the status of issue is closed, we will not skip the test case. Due to the lack of git credentials, we can not visit github and consider all the issues as open before. In this pr, we add the git credentials into environment variables, so we can get the real status of issue. And the status of this issue is closed, we will not skip this test case.

What is the motivation for this PR?
We can use github issues as conditions in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. If the status of issue is closed, we will not skip the test case. Due to the lack of git credentials, we can not visit github and consider all the issues as open before. In this pr, we add the git credentials into environment variables, so we can get the real status of issue. And the status of this issue is closed, we will not skip this test case.

How did you do it?
Set the user name and api token in pipeline/library. Pass them to tests/kvmtest.sh, and export them as environment variables.

How did you verify/test it?
Running the test and the error will not appear again.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>