Added version check to the setup() function (#5589)
The tests within the test_fan_drawer_fans script were all failing because a function call was throwing an exception. On further inspection it was discovered that the function that was being called was an unsupported API in the 201911 and 201811 SONiC OS versions. Code was added to skip those tests if either of these OSes were encountered.