Revert add timestamp to tr.xml and changes in junit xml parser
What is the motivation for this PR?
We found some issues in nightly and smoke test because of adding timestamp and changing in junit xml parser:
1)Some xml report didn't contain "start" but contain "end"
2)Test will fail if xml and json files do not contain "start" and "end"

How did you do it?
Revert changes and fix

How did you verify/test it?
Run a test

Any platform specific information?
Supported testbed topology if it's a new test case?