Get a valid mac adderss for Mellanox switch
The test need to configure a new mac address other than the switch mac on a vlan interface. The community uses "00:11:22:33:44:55" as the new address. But for our asics, we don't support this kind of mac address. So add a new logic to get a valid mac address for Mellanox duts.