[copp] Enhance the copp test case
What is the motivation for this PR?
The CBS is large on Cisco platform, it would take couple of seconds to be exhausted out. If just sending pkt with 10s, the test result is not  accurate because part of them costs on the CBS.

How did you do it?
Change the duration of sending packet from 10s to 30s

How did you verify/test it?
Run copp test.

Signed-off-by: Kevin(Shengkai) Wang <shengkaiwang@microsoft.com>