Modified code to discount PSU voltage value after turn-off
Since each vendor's implementation is different, the PSU's voltage value may be non-zero after shutdown. Additionally, the voltage might be non-zero for a few minutes even after shutdown on account of capacitor physics slowly discharging current and bringing the voltage to zero. Accounting for these discrepancies, and seeing as current flow is a much better indictor of the state of the PSU anyway, voltage value is disregarded in making a decision about the power-state of the PSU.