Remove key "router_macs" in test_decap.py. (#5890)
Description of PR
The key router_macs is not defined in fixture setup_teardown, and it will case an error. In master branch, we no more need this key, so remove it.

What is the motivation for this PR?
The key router_macs is not defined in fixture setup_teardown, and it will case an error. In master branch, we no more need this key, so remove it.

How did you do it?
Remove the key router_macs in ptf_runner in test_decap.py.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>