Optimize DB dump collection for failed test case
The autoused fixture for collecting DB dumps for failed cases need to run
some redis-dump commands for each of the DBs. The original implementation
called the ansible command module on DUT once for each of the redis-dump
command. This is not efficient enough because of the extra ansible overheads.

This change updated the code to use "shell_cmds" module  to run all the
redis-dump commands in a batch.

Tested on vSONiC setup, this change can decrease redis-dump execution
time from 13 seconds to 6 seconds.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>