[GCU] Improve GCU test from invalid initial config (#5886)
Summary: Improve GCU nightly test when initial config is invalid

What is the motivation for this PR?
GCU required the ConfigDB to pass YANG validation. But previous failing test may leave invalid residue to ConfigDB.
This PR is to make GCU test coud be tested in this case.

How did you do it?
Check initial ConfigDB and reload with minigraph if it is invalid.

How did you verify/test it?
Run kvm test with config with YANG validation failed.