Address the security alert of aiohttp used by test report uploading (#5903)
What is the motivation for this PR?
Github is giving security vulnerabilities alert for the aiohttp
package used by test report uploading.

How did you do it?
This change addressed the security alert by removing the dependency
of this aiohttp package by upgrading the packages depending on it
to a higher version.

The pytest package is only required for testing. It is not required
for test report uploading. Moved it into a separate rquirements_dev.txt
file. The README.md file was also updated.

Currently the test report uploading tests are broken. Need to address
them in separate PR.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>