Merged PR 678: Get token from environment variable for accessing Github
We can use github issues as conditions in `tests/common/plugins/conditional_mark/tests_mark_conditions.yaml`. If the status of issue is `closed`, we will not skip the test case. Due to the lack of git credentials, we can not visit github and consider all the issues as open before. In this pr, we add the git credentials into environment variables, so we can get the real status of issue. And the status of this issue is closed, we will not skip this test case.