Optimize DB dump collection for failed test case (#5910)
What is the motivation for this PR?
The autoused fixture for collecting DB dumps for failed cases need to run
some redis-dump commands for each of the DBs. The original implementation
called the ansible command module on DUT once for each of the redis-dump
command. This is not efficient enough because of the extra ansible overheads.

How did you do it?
This change updated the code to use "shell_cmds" module to run all the
redis-dump commands in a batch.
This change also updated the dump file name to include test module name,
not just the test case name which could conflict across different test scripts.

How did you verify/test it?
Tested on vSONiC setup, this change can decrease redis-dump execution
time from 13 seconds to 6 seconds.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>