Skip generic_config_updater tests for T2 topology (#5898)
What is the motivation for this PR?
As per discussion with MSFT, the tests in generic_config_updater use YANG models which are not currently supported on a T2 chassis. So, need to skip these tests for T2 topology

How did you do it?
Adding check in tests_mark_conditional to skip generic_config_updater if 't2' in topology

How did you verify/test it?
Ran generic_config_updater against T2 topology and verified that all the tests are skipped.