[Mellanox] Use the correct way to check branch in "check_sysfs" (#5913)
Use the correct way to check the branch of DUT in "check_sysfs"

- What is the motivation for this PR?
Use the correct way to check the release of DUT in "check_sysfs"
dut.os_version is generated from the branch name. There is no guarantee that branch name always includes the release name. For example, a developer can create a branch my-fix based on 202012 branch. In this case, the test will fail.

- How did you do it?

- How did you verify/test it?
Run sonic-mgmt test which noticed this issue

- Any platform specific information?
Mellanox

Signed-off-by: Stephen Sun <stephens@nvidia.com>