Cache dut basic fact using `session.config.cache`.  (#5932)
Description of PR
In order to skip test cases using conditional marks, we need to gather the basic facts of dut, which will cost some time. session.config.cache in pytest can support sharing values between py.test invocations. So we use this to cache the basic facts of dut.

What is the motivation for this PR?
In order to skip test cases using conditional marks, we need to gather the basic facts of dut, which will cost some time. session.config.cache in pytest can support sharing values between py.test invocations. So we use this to cache the basic facts of dut.

How did you do it?
Compare the tbname with cached tbname.
If they are not the same testbed, clear the cache, get the new basic facts, and put the new tbname and basic facts in cache.
If the tbname is same as cached tbname, if we already have dut basic facts in cache, use them ditectly, otherwise, get the basic facts and put them in cache.

How did you verify/test it?
Running test cases with/without cached tbname and dut basic facts.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>