[GCU] Improve loopback test in dualTor (#5933)
Summary: This PR is to improve lo test to fit dualTor scenarios.

What is the motivation for this PR?
In dualTor scenarios, one Tor's Loopback0 is 10.1.0.33 which is not same as the default 10.1.0.32. Need to improve test to fit in that.

How did you do it?
Read loopback config from cfg_facts.

How did you verify/test it?
Run kvm test.