Add a check for PCIE_DEVICES table status before getting the key
Before getting the key of PCIE_DEVICE table, check if the table is ready. The reason is if the previous test has a config reload in teardown, even if the critical processes are all running, it still takes some time to restore all the data tables including the PCIE_DEVICE. There is a chance the table is still empty when the pcie test starts to run and try to get the key.