Merged PR 722: Explicitly set job timeout value to override the default 60 minutes
Explicitly set job timeout value to override the default 60 minutes

PR #710 removed the job timeout value and added timeout values to tasks that
may be stuck. The problem is that when no timeout is specified for a job,
it will have default 60 minutes timeout value. This is too short to meet
nightly test needs.

If set timeout value to 0 for the nightly test job, will may run forever
on self-hosted agent. To avoid potential issue, this fix set the timeout
value of the nightly test job to 2400 minutes (40 hours). This value should
be long enough.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>