support docker registry (#5929)
What is the motivation for this PR?
For the QoS test, need to swap syncd during run time, sometimes it is necessary to use the other docker registry for different image versions. In current framework, we can specify the docker registry host per inventory but not per image.

In this PR, to add a new yml file which can give an option for a different docker registry only for the QoS swap syncd.

How did you do it?
How did you verify/test it?
PASSED the nightly test to do the swap syncd with public image

Any platform specific information?
Supported testbed topology if it's a new test case?