Updated approach for installing nanomsg in copp TC (#5962)
Summary:

Updated checking availability of the nanomsg in the syncd container
Added check Debian version before installation of nano
Added approach for disabling bgp before installation of nano

What is the motivation for this PR?
syncd container was upgraded to Debian 11 Bullseye. test_copp uses Python 2 to install nano in syncd container, but Debian 11 doesn't support Python 2. #5905 PR of sonic-mgmt added approach for installing the correct version of Python, but this doesn't cover the case when python isn't available. So that approach was updated.

How did you do it?
Updated approach for installing nanomsg in copp TC

Signed-off-by: Oleksandr Kozodoi <oleksandrx.kozodoi@intel.com>