Merged PR 731: Schedule smoke test to run sooner after merge is done
Schedule smoke test to run sooner after merge is done

Currently the smoke test is scheduled to 12:00 UTC time. The merge
is scheduled on 02:30. There is some unnecessary wait time between
merge and smoke test.

This change updated the scedule time of smoke test so that it can
start as soon as merge is completed.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>