Merged PR 751: Skip sub_port_interfaces for other asics except td2 and skip it for 201811 an...
Skip sub_port_interfaces for other asics except td2 and skip it for 201811 and 201911.

Internally, we only need to run sub_port_interfaces testcases on TD2 (7050qx platform) for 202012 and above. We can skip all other asics.

For avoiding the conflict in githut master, update it in the internal branch to override the one used in public github.

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>