[test_route_perf] increase route change check timeout (#5940)
Summary:
The Dynamic Inner Hashing tests were sensitive to the order in a teardown.
Sometimes the order of fixtures was wrong. This led to the failure of the next tests.
To add stability were changed the setup and teardown methods were. To use the functions instead of fixtures and config reload for all system cleaning.

What is the motivation for this PR?
stabilize the test

How did you do it?
changed the setup and teardown methods

Signed-off-by: Anton <antonh@nvidia.com>