Add a check for PCIE_DEVICES table status before getting the key (#5944)
Before getting the key of PCIE_DEVICE table, check if the table is ready. The reason is if the previous test has a config reload in teardown, even if the critical processes are all running, it still takes some time to restore all the data tables including the PCIE_DEVICE. There is a chance the table is still empty when the pcie test starts to run and try to get the key.

- What is the motivation for this PR?
Add a check for PCIE_DEVICES table status before getting the key to improve test stability.

- How did you do it?
Add a function to check the PCIE_DEVICES table status, and wait until the table is not empty for at most 30 seconds.

- How did you verify/test it?
Run regression test on all Nvidia platforms with 202012 and 202205 version. All passed.