enum_asic fixtures to use only asics that are present instead of all asics based on num_asics (#5828)
What is the motivation for this PR?
Tests that are parameterized using the enum_asic fixtures fail when the selected asic index is not present (operational) in the DUT.

Example of this is on the supervisor card in a chassis, where we could have not all SFMs in the chassis. In this case, only asics
corresponding to the SFMs present in the chassis are operational, but not the others. For example, on a Nokia chassis, if we have only 3 SFM's then
only 6 (2 per SFM) out of the possible max 16 asics are operational.

How did you do it?
Added optional "asics_present" to the inventory data of the DUT in the inventory file. This represents the list of asics that are operational on the DUT.
If asics_present is present in the inventory data, we would use that list for generating the list of asics to use. Else, we use range of 'num_asics' defined in the inventory data to generate the list of asics to use.
Example inventory data for a supervisor card

supervisor1:
num_asics: 16
asics_present: [12,13]

In the above case, the asic list will be only 12 and 13 and not 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15

How did you verify/test it?
Tested the code on a dut that did not have all fabric-cards for the following fixtures:

enum_dut_hostname
enum_asic_index
enum_frontend_asic_index
enum_backend_asic_index
enum_rand_one_asic_index
enum_rand_one_frontend_asic_index
Ran autorestart container test against the same chassis - which used enum_rand_one_asic_index and it picked the asic that was operational.