Improve robustness of remove-topo operation
The `testbed-cli.sh remove-topo` operation is to remove the topology, including cEOS neighbors, ovs bridge bindings, and ovs bridges.

For some reason, the ovs bridges to be removed may be gone. In this case, the `testbed-cli.sh remove-topo` will fail with not able to find the bridges.

Indeed this failure is unnecessary because this operation is to remove the bridge bindings and bridges. If a bridge is already gone, then we can simply ignore it.

This change improved the vm_topology ansible module to skip a bridge if it does not exit while trying to remove the bridge bindings and bridges. With this change, the `testbed-cli.sh remove-topo` always can succeed and do the job.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>