Merged PR 714: To support public docker registry for the swap syncd which is required by test_qos_sai
What is the motivation for this PR?
There are different docker registry for the public and internal images, for security reason, we can't upload public and internal images into same docker registry, but this is needed for the tst_qos_sai testcases.

How did you do it?
In current frame work, it is not supported to have different docker registry per image. In this PR, to add a new yml file which can offer a new option for the public image to have a different docker registry. And since all public images will share one docker registry host, to add the support in the nightly_test.yml to identify the public image would be good enough.

How did you verify/test it?
PASSED the nightly test to do the swap syncd with public image
https://dev.azure.com/mssonic/internal/_build/results?buildId=118511&view=results

Any platform specific information?
Non marvell