[dualtor][active-active] Unblock fib tests (#6006)
Approach
What is the motivation for this PR?
Enable test_fib on dualtor-mixed testbeds.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
1. Add fixture mux_status_from_nic_simulator to interacts with nic_simulator to retrieve mux status for ports in active-active cable type.
2. Add fixture ptf_test_port_map_active_active to build ptf port map that supports dualtor-mixed testbed.
The key difference is that, for packets ingressing ptf ports connected to DUTs' active-active cable type ports, the target DUT could be either upper ToR or lower ToR(both ToRs are active), the generated ptf port mapping will be like:
``` 
u'3': {u'asic_idx': 0,
        u'target_dest_mac': u'00:aa:bb:cc:dd:ee',
        u'target_dut': [0, 1],
        u'target_src_mac': [u'2c:dd:e9:0f:4e:50', u'2c:dd:e9:0f:3d:4c']}
```
3. Enable ptftests hash_test and fib_test to do I/O verification for this multi-next-hop scenario.
For a packet sent from ptf port that connects to an active-active port, the test will try to use the fibs from each ToR to parse its nexthops, and verify the packet forwarding on those nexthops. For ECMP behavior, the test will only check balancing over a single ToR.

For example, if a packet destinated to 1.1.1.1 is sent to ptf port eth3, which is connected Ethernet12 of both ToRs, the test will determine the nexthops on both ToRs, if both ToRs forward this packet with the default route, and upper ToR will forward this packet to ptf ports [30, 32, 34, 36], and lower ToR will forward this packet to ptf ports [31, 33, 35, 37], the test will try to verify the packet forwarding on ports [30, 31, 32, 33, 34, 35, 36, 37]. And for balancing, the test will verify the traffic is balanced over a single ToR, so it will try to verify balancing on ports set [30, 32, 34, 36] or [31, 33, 35, 37] separately.

How did you verify/test it?
Run test_fib on dualtor, dualtor-mixed, t0, and t1

Any platform specific information?
Supported testbed topology if it's a new test case?