Skipping test for absent psu's (#5679)
What is the motivation for this PR?
Currently the tests are trying to access PSU which are not present and failing for them. Even though there is a skip_psu_list, there is no code present to use it. Added the support for it in all the relevant cases.

How did you do it?

How did you verify/test it?
Verified against cisco-8000 platform

Co-authored-by: Perumal Venkatesh <pevenkat@cisco.com>