Merged PR 778: Try to redeploy topology if restart-ptf failed in nightly test
Try to redeploy topology if restart-ptf failed in nightly test

Recently we frequently hit the issue of restart-ptf failure in nightly test.
The reason is that some testbed topology was broken. The cEOS docker
containers and their associated ovs bridges are gone. When this issue happens,
'testbed-cli.sh remove-topo' will fail too.
PR https://github.com/Azure/sonic-mgmt/pull/6020 improved the robustness of
'testbed-cli.sh remove-topo' for such scenario. With the help of that change,
then we can improve the nightly test to run a 'remove-topo' and 'add-topo'
if 'restart-ptf' failed.

PR https://github.com/Azure/sonic-mgmt/pull/5945 introduced a new testbed-cli.sh
command 'redeploy-topo' which is a wrapper for 'remove-topo' and 'add-topo'.

So, in this change, we can run 'testbed-cli.sh redeploy-topo' if restart-ptf failed.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>