Merged PR 786: Add validation for vendor password in kickstart.
What is the motivation for this PR?
When password is null for vendor devices, the device does not accept the invalid password and is waiting for valid password. The waiting process will take a long time until an error is generated.

How did you do it?
Before starting vendor VMs, check whether password is null, if password is null then prompt an error and stop the script.

How did you verify/test it?
Run the testcase with vendor password to '' and '   ', the script can stop before vendor device starts.
Run the testcase with correct format of vendor password, script can run successfully.