Disable CI testing keep only PR testing (#6061)
What is the motivation for this PR?
Currently the pipeline has both PR and CI testing enabled. After a PR is merged,
the pipeline is triggered again. This is a waste of resource.

How did you do it?
This change disabled CI testing. The PR testing will be still triggered as usual.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>