Fix ceos duplicate mac address issue on Ubuntu 22.04
If deploy a topology using cEOS, one of the step is to create veth interfaces
for the cEOS docker containers.

For example, the current steps to create backplane interfaces:

1.1 Create veth pair in host for container VM0100
      ip link add VM0100-back type veth peer name eth5
1.2 Add the eth5 interface to network namespace of container VM0100

2.1 Create veth pair in host for another container VM0101
      ip link add VM0101-back type veth peer name eth5
2.2 Add the eth5 interface to network namespace of container VM0101

As we can see that after step 1.2, eth5 is no longer in the host namespace.
Then in step 2.1 we can add another interface with same name eth5.

The problem is that on Ubuntu 22.04, mac address of eth5 created in step 2.1
will be the same as the eth5 interface created in step 1.1. Possibly
Ubuntu 22.04 is using a different algorithm for assigning mac address
to new veth interfaces. If interface name is same, then mac address
will be same too.

Because all the VMxxxx-back interfaces will be attached to a same
ovs bridge, their peer interfaces should not use same mac address.

The fix is to create veth interfaces with unique name for all cEOS
containers.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>