Add param to let user update the skip reason
In the skip mark condition file, user can specify the skip reason, but sometimes user may need to use the skip reason to do some deep process, then the skip reason is not realiable since user can specify any string in the skip reason, but the condition in the skip mark condition file need to follow some rule, so we can update the skip reason with the condition, and it will be easier for user to process the skip reason. if the --dynamic_update_skip_reason is not specified in the pytest, it will not update the skip reason, so the default behavior of this file will not be changed.