Replace `pytest.skip` in test scripts with conditional marks.  (#5708)
Description of PR
When we use pytest.skip in test scripts, it will first execute some fixtures in test cases, which will waste some time. When using conditional marks to skip test cases, it will skip the case in the collect period, which will not execute fixtures in test cases and save some execute time. In this pr, we replace pytest.skip in test scripts with conditional marks to skip test case in advance and save execute time.

What is the motivation for this PR?
When we use pytest.skip in test scripts, it will first execute some fixtures in test cases, which will waste some time. When using conditional marks to skip test cases, it will skip the case in the collect period, which will not execute fixtures in test cases and save some running time. In this pr, we replace pytest.skip in test scripts with conditional marks to skip test case in advance and save running time.

How did you do it?
Replace pytest.skip in test scripts with conditional marks in tests_mark_conditions.yaml.

How did you verify/test it?
By running whole test cases and observe the running time.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>